refactor: separate parsing from database managing
